requirements.txt
requests
python-dotenv
schedule
Flask
selenium
